ssome norminator action
