+++ src/inventory/inventory_equip.c
@@ -6,7 +6,7 @@
 /*   By: ahakanen <aleksi.hakanen94@gmail.com>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/08 22:33:29 by ahakanen          #+#    #+#             */
-/*   Updated: 2021/04/05 17:59:46 by ahakanen         ###   ########.fr       */
+/*   Updated: 2021/04/07 14:19:17 by ahakanen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
