+++ src/inventory/inventory_init_items.c
@@ -6,7 +6,7 @@
 /*   By: ahakanen <aleksi.hakanen94@gmail.com>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/11 11:15:29 by ahakanen          #+#    #+#             */
-/*   Updated: 2021/04/05 17:59:46 by ahakanen         ###   ########.fr       */
+/*   Updated: 2021/04/07 14:19:42 by ahakanen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
