+++ src/inventory/item_data/item_data_pistol.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   item_data_pistol.c                                  :+:      :+:    :+:   */
+/*   item_data_pistol.c                                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: ahakanen <aleksi.hakanen94@gmail.com>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2021/01/11 12:04:24 by ahakanen          #+#    #+#             */
-/*   Updated: 2021/04/02 15:09:17 by ahakanen         ###   ########.fr       */
+/*   Created: 2021/04/07 14:12:35 by ahakanen          #+#    #+#             */
+/*   Updated: 2021/04/07 14:12:46 by ahakanen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
