+++ src/npc/npc_animation.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   npc_animation.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ohakola <ohakola@student.hive.fi>          +#+  +:+       +#+        */
+/*   By: ahakanen <aleksi.hakanen94@gmail.com>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/12/30 18:41:09 by veilo             #+#    #+#             */
-/*   Updated: 2021/02/15 21:41:51 by ohakola          ###   ########.fr       */
+/*   Updated: 2021/04/07 14:22:09 by ahakanen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
