+++ src/projectile/projectile_init.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   projectile_init.c                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ohakola <ohakola@student.hive.fi>          +#+  +:+       +#+        */
+/*   By: ahakanen <aleksi.hakanen94@gmail.com>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/11 13:16:45 by ahakanen          #+#    #+#             */
-/*   Updated: 2021/01/12 21:41:44 by ohakola          ###   ########.fr       */
+/*   Updated: 2021/04/07 14:36:34 by ahakanen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
